Lowercase directory names; more metadata to summaries renames
