+++ frontend/src/App.tsx
@@ -7,7 +7,7 @@ import AppLayout from "./components/app-layout/AppLayout";
 import { CssBaseline, ThemeProvider } from "@mui/material";
 import { appTheme } from "./components/common/styles/theme";
 import Leaderboard from "./pages/leaderboard/Leaderboard";
-import Home from "./pages/Home/Home";
+import Home from "./pages/home/Home";
 import News from "./pages/news/News";
 
 function App() {
